% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disastr.api.R
\name{disastr.api}
\alias{disastr.api}
\title{Automated Retrieval of Disaster Event Data}
\usage{
disastr.api(limit = NULL, ascending = FALSE, other.query = NULL)
}
\arguments{
\item{limit}{numeric. Provide the number of events you wish to retrieve. When set to NULL (default), ten events are returned.}

\item{ascending}{logical. If set to FALSE (default), events will be returned in decreasing order starting from the most
recent. If set to TRUE, events will be returned in ascending order starting from the oldest record.}

\item{other.query}{character vector. Allows users to add their own ReliefWeb API queries to the
GET call. See the \href{https://apidoc.rwlabs.org/}{ReliefWeb documentation} for query details. Vector elements are
assumed to be individual queries, and are automatically separated by an & sign.}
}
\value{
A data frame containing ReliefWeb disaster events.
}
\description{
This small package provides functionality to access and manage the application programming
interface (API) of the \href{https://reliefweb.int/disasters/}{United Nations Office for the Coordination of Humanitarian Affairs' (OCHA) ReliefWeb disaster events}.
The function \emph{\code{disastr.api()}} makes it easy to retrieve a user-defined sample (or all of the
available data) from ReliefWeb, enabling a seamless integration of regular data updates into
the research work flow.
}
\details{
The function \emph{\code{disastr.api()}} is an R wrapper for
the \href{https://reliefweb.int/disasters/}{United Nations Office for the Coordination of Humanitarian Affairs' (OCHA) ReliefWeb disaster events} API.
Internally it uses \emph{\code{httr}} to access the API, and \emph{\code{jsonlite}} to manage the JSON content that the call returns. The JSON data
are converted into the base class \emph{\code{data.frame}}. Variables are of class \emph{\code{character}} by default.
Variables which only contain numbers as recognized by the regular
expression \verb{^[0-9]+$} are coerced into \emph{\code{numeric}} before the \emph{\code{data.frame}} object is returned. \cr \cr
}
\examples{
my.data.frame1 <- disastr.api()
head(my.data.frame1)

}
\references{
United Nations Office for the Coordination of Humanitarian Affairs. \emph{ReliefWeb Disasters}. \url{https://reliefweb.int/disasters/} \cr
}
\author{
Christoph Dworschak \cr Website: \href{https://www.chrisdworschak.com/}{\url{https://chrisdworschak.com/}}
}
